requirements.txt
python-telegram-bot
Flask
uvicorn
asgiref
pathos
requests